+++ src/args.rs
@@ -882,6 +882,7 @@ impl ArgMatches {
             .git_global(!self.no_ignore_vcs() && !self.no_ignore_global())
             .git_ignore(!self.no_ignore_vcs())
             .git_exclude(!self.no_ignore_vcs() && !self.no_ignore_exclude())
+            .require_git(!self.is_present("no-require-git"))
             .ignore_case_insensitive(self.ignore_file_case_insensitive());
         if !self.no_ignore() {
             builder.add_custom_ignore_filename(".rgignore");